feat(core): implement support for auth v3
Auth v3 is an extension to the existing auth v2 authentication scheme,
which adds defenses against replay attacks and protocol downgrades.

This commit adds support for Auth v3 to the SDK, and adds flags for
enabling mandatory auth v3 to BitGo Express. Auth v2 is still the
default auth scheme for now until we gain more confidence that Auth v3
is working as intended.

Ticket: BG-30091